+++ src/uhttp/io/FileList.cpp
@@ -18,8 +18,8 @@ using namespace uHTTP;
 ////////////////////////////////////////////////
 
 void FileList::clear()  {
-  int nNode = size();
-  for (int n = 0; n < nNode; n++) {
+  size_t nNode = size();
+  for (size_t n = 0; n < nNode; n++) {
     File *file = getFile(n);
     delete file;
   }
